split response console tests
