+++ x-pack/plugins/security_solution/public/management/cypress/e2e/response_actions/response_console/execute.cy.ts
@@ -1,0 +1,74 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+import type { PolicyData } from '../../../../../../common/endpoint/types';
+import type { CreateAndEnrollEndpointHostResponse } from '../../../../../../scripts/endpoint/common/endpoint_host_services';
+import {
+  inputConsoleCommand,
+  openResponseConsoleFromEndpointList,
+  submitCommand,
+  waitForCommandToBeExecuted,
+  waitForEndpointListPageToBeLoaded,
+} from '../../../tasks/response_console';
+import type { IndexedFleetEndpointPolicyResponse } from '../../../../../../common/endpoint/data_loaders/index_fleet_endpoint_policy';
+import { createAgentPolicyTask, getEndpointIntegrationVersion } from '../../../tasks/fleet';
+
+import { login } from '../../../tasks/login';
+import { enableAllPolicyProtections } from '../../../tasks/endpoint_policy';
+import { createEndpointHost } from '../../../tasks/create_endpoint_host';
+import { deleteAllLoadedEndpointData } from '../../../tasks/delete_all_endpoint_data';
+
+describe('Response console', { tags: ['@ess', '@serverless'] }, () => {
+  beforeEach(() => {
+    login();
+  });
+
+  describe('Execute operations: execute', () => {
+    const homeFilePath = process.env.CI || true ? '/home/vagrant' : `/home/ubuntu`;
+
+    let indexedPolicy: IndexedFleetEndpointPolicyResponse;
+    let policy: PolicyData;
+    let createdHost: CreateAndEnrollEndpointHostResponse;
+
+    before(() => {
+      getEndpointIntegrationVersion().then((version) =>
+        createAgentPolicyTask(version).then((data) => {
+          indexedPolicy = data;
+          policy = indexedPolicy.integrationPolicies[0];
+
+          return enableAllPolicyProtections(policy.id).then(() => {
+            // Create and enroll a new Endpoint host
+            return createEndpointHost(policy.policy_id).then((host) => {
+              createdHost = host as CreateAndEnrollEndpointHostResponse;
+            });
+          });
+        })
+      );
+    });
+
+    after(() => {
+      if (createdHost) {
+        cy.task('destroyEndpointHost', createdHost);
+      }
+
+      if (indexedPolicy) {
+        cy.task('deleteIndexedFleetEndpointPolicies', indexedPolicy);
+      }
+
+      if (createdHost) {
+        deleteAllLoadedEndpointData({ endpointAgentIds: [createdHost.agentId] });
+      }
+    });
+
+    it('"execute --command" - should execute a command', () => {
+      waitForEndpointListPageToBeLoaded(createdHost.hostname);
+      openResponseConsoleFromEndpointList();
+      inputConsoleCommand(`execute --command "ls -al ${homeFilePath}"`);
+      submitCommand();
+      waitForCommandToBeExecuted('execute');
+    });
+  });
+});
