+++ x-pack/plugins/security_solution/public/management/cypress/e2e/response_actions/response_console/process_operations.cy.ts
@@ -1,0 +1,126 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { PolicyData } from '../../../../../../common/endpoint/types';
+import type { CreateAndEnrollEndpointHostResponse } from '../../../../../../scripts/endpoint/common/endpoint_host_services';
+import {
+  inputConsoleCommand,
+  openResponseConsoleFromEndpointList,
+  performCommandInputChecks,
+  submitCommand,
+  waitForCommandToBeExecuted,
+  waitForEndpointListPageToBeLoaded,
+} from '../../../tasks/response_console';
+import type { IndexedFleetEndpointPolicyResponse } from '../../../../../../common/endpoint/data_loaders/index_fleet_endpoint_policy';
+import { createAgentPolicyTask, getEndpointIntegrationVersion } from '../../../tasks/fleet';
+
+import { login } from '../../../tasks/login';
+import { enableAllPolicyProtections } from '../../../tasks/endpoint_policy';
+import { createEndpointHost } from '../../../tasks/create_endpoint_host';
+import { deleteAllLoadedEndpointData } from '../../../tasks/delete_all_endpoint_data';
+
+describe('Response console', { tags: ['@ess', '@serverless'] }, () => {
+  beforeEach(() => {
+    login();
+  });
+
+  describe('Processes operations: list, kill and suspend process', () => {
+    let cronPID: string;
+    let newCronPID: string;
+
+    let indexedPolicy: IndexedFleetEndpointPolicyResponse;
+    let policy: PolicyData;
+    let createdHost: CreateAndEnrollEndpointHostResponse;
+
+    before(() => {
+      getEndpointIntegrationVersion().then((version) =>
+        createAgentPolicyTask(version).then((data) => {
+          indexedPolicy = data;
+          policy = indexedPolicy.integrationPolicies[0];
+
+          return enableAllPolicyProtections(policy.id).then(() => {
+            // Create and enroll a new Endpoint host
+            return createEndpointHost(policy.policy_id).then((host) => {
+              createdHost = host as CreateAndEnrollEndpointHostResponse;
+            });
+          });
+        })
+      );
+    });
+
+    after(() => {
+      if (createdHost) {
+        cy.task('destroyEndpointHost', createdHost);
+      }
+
+      if (indexedPolicy) {
+        cy.task('deleteIndexedFleetEndpointPolicies', indexedPolicy);
+      }
+
+      if (createdHost) {
+        deleteAllLoadedEndpointData({ endpointAgentIds: [createdHost.agentId] });
+      }
+    });
+
+    it('"processes" - should obtain a list of processes', () => {
+      waitForEndpointListPageToBeLoaded(createdHost.hostname);
+      openResponseConsoleFromEndpointList();
+      performCommandInputChecks('processes');
+      submitCommand();
+      cy.contains('Action pending.').should('exist');
+      cy.getByTestSubj('getProcessesSuccessCallout', { timeout: 120000 }).within(() => {
+        ['USER', 'PID', 'ENTITY ID', 'COMMAND'].forEach((header) => {
+          cy.contains(header);
+        });
+
+        cy.get('tbody > tr').should('have.length.greaterThan', 0);
+        cy.get('tbody > tr > td').should('contain', '/usr/sbin/cron');
+        cy.get('tbody > tr > td')
+          .contains('/usr/sbin/cron')
+          .parents('td')
+          .siblings('td')
+          .eq(1)
+          .find('span')
+          .then((span) => {
+            cronPID = span.text();
+          });
+      });
+    });
+
+    it('"kill-process --pid" - should kill a process', () => {
+      waitForEndpointListPageToBeLoaded(createdHost.hostname);
+      openResponseConsoleFromEndpointList();
+      inputConsoleCommand(`kill-process --pid ${cronPID}`);
+      submitCommand();
+      waitForCommandToBeExecuted('kill-process');
+
+      performCommandInputChecks('processes');
+      submitCommand();
+
+      cy.getByTestSubj('getProcessesSuccessCallout', { timeout: 120000 }).within(() => {
+        cy.get('tbody > tr > td')
+          .contains('/usr/sbin/cron')
+          .parents('td')
+          .siblings('td')
+          .eq(1)
+          .find('span')
+          .then((span) => {
+            newCronPID = span.text();
+          });
+      });
+      expect(newCronPID).to.not.equal(cronPID);
+    });
+
+    it('"suspend-process --pid" - should suspend a process', () => {
+      waitForEndpointListPageToBeLoaded(createdHost.hostname);
+      openResponseConsoleFromEndpointList();
+      inputConsoleCommand(`suspend-process --pid ${newCronPID}`);
+      submitCommand();
+      waitForCommandToBeExecuted('suspend-process');
+    });
+  });
+});
